+++ CHANGELOG.md
@@ -11,8 +11,8 @@
  - Added Product Batch controller with error handling and response formatting.
  - Created routes for Product Batch operations with JSDoc documentation.
  - UUID generation for inventory IDs.
+ - Added new fields for workshops model and controller.
  
-
 ## [0.0.0] - 04/11/2025
 
 ### Fixed